Add authorization, authentication, validation, finish app

+++ controllers/user.js
@@ -1,49 +1,12 @@
 // Import module
 const bcrypt = require("bcrypt");
+const jwt = require("jsonwebtoken");
 
 // Import models
 const { User } = require("../models");
 const { Favorite_Track } = require("../models");
 const { Followed_Artist } = require("../models");
 
-const allUsers = async (req,res) => {
-    try {
-        const users = await User.findAll({
-            attributes: ["id", "name", "email"]
-        });
-        res.status(200).json({
-            message: "All Users",
-            data: users
-        });
-    } catch (error) {
-        res.status(404).json({
-            message: error.message
-        });
-    }
-}
-
-// Have to add authorization
-const userProfile = async (req,res) => {
-    try {
-        const userId = req.params.userId;
-        const profile = await User.findOne({
-            where: {
-                id: userId
-            },
-            attributes: ["id", "name", "email"]
-        });
-        res.status(200).json({
-            message: "User Profile",
-            data: profile
-        });
-    } catch (error) {
-        res.status(404).json({
-            message: error.message
-        });
-    }
-}
-
-// Add authentication (check whether the email is available) & validation of the password - email requirements
 const registerUser = async (req,res) => {
     let salt = bcrypt.genSaltSync(10);
     const body = req.body;
@@ -56,27 +19,125 @@ const registerUser = async (req,res) => {
         email: email, 
         password: password
     };
-    try {
-        await User.create(user);
-        res.status(200).json({
-            message: "Register New User",
-            data: {
-                name: name,
-                email: email
-            }
-        });
-    } catch (error) {
-        res.status(404).json({
-            message: error.message
-        });
+    if(body.password === passwordConfirm){
+        try {
+            await User.create(user);
+            res.status(200).json({
+                message: "Register New User",
+                data: {
+                    name: name,
+                    email: email
+                }
+            });
+        } catch (error) {
+            res.status(400).json({
+                message: error.message
+            });
+        }
+    } else {
+        res.status(400).json({
+            message: "Password & password confirmation contain different values."
+        })
     }
+    
 }
 
 // Login
+const loginUser = async (req,res) => {
+    const body = req.body;
+    const email = body.email;
+    const password = body.password;
+    // Check if the user exists
+    const user = await User.findOne({
+        where: {
+            email: email,
+        }
+    });
 
-// Add authorization
+    // If user exists, send JWT as response, else error
+    if(user && bcrypt.compareSync(password, user.password)){
+        // Admin
+        if(email === "aurellia_christie@yahoo.com"){
+            const token = jwt.sign({
+                id: user.id,
+                email: user.email,
+                role: "admin"
+            },
+                process.env.TOKEN
+            )
+            res.status(200).json({ 
+                message: "Logged In.",
+                token: token
+             });
+        } else {
+        // User
+            const token = jwt.sign({
+                id: user.id,
+                email: user.email,
+                role: "user"
+            },
+                process.env.TOKEN
+            )
+            res.status(200).json({ 
+                message: "Logged In.",
+                token: token
+             });
+        }       
+    }else{
+        res.status(401).send("Email or password is incorrect.");
+    }
+} 
+
+const allUsers = async (req,res) => {
+    if(req.user.role === "admin"){
+        try {
+            const users = await User.findAll({
+                attributes: ["id", "name", "email"]
+            });
+            res.status(200).json({
+                message: "All Users",
+                data: users
+            });
+        } catch (error) {
+            res.status(404).json({
+                message: error.message
+            });
+        }
+    } else {
+        res.status(403).json({
+            message: "Forbidden User."
+        })
+    }   
+}
+
+const userProfile = async (req,res) => {
+    const userId = Number(req.params.userId);
+    if(req.user.id === userId){
+        try {
+            const profile = await User.findOne({
+                where: {
+                    id: userId
+                },
+                attributes: ["id", "name", "email"]
+            });
+            res.status(200).json({
+                message: "User Profile",
+                data: profile
+            });
+        } catch (error) {
+            res.status(404).json({
+                message: error.message
+            });
+        }
+    } else {
+        res.status(403).json({
+            message: "Forbidden User."
+        })
+    }
+}
+
 const editUserProfile = async (req,res) => {
-    const userId = req.params.userId;
+    const userId = Number(req.params.userId);
     const body = req.body;
     const name = body.name;
     const email = body.email;
@@ -84,53 +145,63 @@ const editUserProfile = async (req,res) => {
         name: name, 
         email: email, 
     };
-    try {
-        await User.update(profile, {
-            where: {
-                id: userId
-            }
-        });
-        res.status(200).json({
-            message: "Update User Profile",
-            data: {
-                name: name,
-                email: email
-            }
-        });
-    } catch (error) {
-        res.status(404).json({
-            message: error.message
-        });
+    if(req.user.id === userId){
+        try {
+            await User.update(profile, {
+                where: {
+                    id: userId
+                }
+            });
+            res.status(200).json({
+                message: "Update User Profile",
+                data: {
+                    name: name,
+                    email: email
+                }
+            });
+        } catch (error) {
+            res.status(404).json({
+                message: error.message
+            });
+        }
+    } else {
+        res.status(403).json({
+            message: "Forbidden User."
+        })
     }
 }
 
-// Add authorization
 const deleteUser = async (req,res) => {
-    const userId = req.params.userId; 
-    try {
-        await User.destroy({
-            where: {
-                id: userId
-            }
-        });
-        res.status(200).send({
-            message: `Delete User.`
-        });
-    } catch (error) {
-        res.status(404).json({
-            message: error.message
-        });
+    const userId = Number(req.params.userId); 
+    if(req.user.id === userId || req.user.role === "admin"){
+        try {
+            await User.destroy({
+                where: {
+                    id: userId
+                }
+            });
+            res.status(200).send({
+                message: `Delete User.`
+            });
+        } catch (error) {
+            res.status(404).json({
+                message: error.message
+            });
+        }
+    } else{
+        res.status(403).json({
+            message: "Forbidden User."
+        })
     }
 }
 
-// Add authorization
 const addFavoriteTrack = async (req,res) => {
     const body = req.body;
     const title = body.title;
     const cover = body.cover;
     const artist = body.artist;
     const preview = body.preview;
-    const userId = req.params.userId;
+    const userId = Number(req.params.userId);
     const track = {
         title: title,
         cover: cover,
@@ -138,128 +209,152 @@ const addFavoriteTrack = async (req,res) => {
         preview: preview,
         user_id: userId
     };
-    try {
-        await Favorite_Track.create(track);
-        res.status(200).json({
-            message: "Add Favorite Track",
-            data: {
-                track
-            }
-        });
-    } catch (error) {
-        res.status(404).json({
-            message: error.message
-        });
+    
+    if(req.user.id === userId){
+        try {
+            await Favorite_Track.create(track);
+            res.status(200).json({
+                message: "Add Favorite Track",
+                data: track
+            });
+        } catch (error) {
+            res.status(404).json({
+                message: error.message
+            });
+        }
+    } else {
+        res.status(403).json({
+            message: "Forbidden User."
+        })
     }
 }
 
-// Add authorization
 const favoriteTracks = async (req,res) => {
-    const userId = req.params.userId;
-    try {
-        const tracks = await Favorite_Track.findAll({
-            where: {
-                user_id: userId
-            },
-            attributes: ['title', 'cover', 'artist', 'preview']
-        });
-        res.status(200).json({
-            message: "Favorite Tracks",
-            data: tracks
-        });
-    } catch (error) {
-        res.status(404).json({
-            message: error.message
-        });
+    const userId = Number(req.params.userId);
+    if(req.user.id === userId){
+        try {
+            const tracks = await Favorite_Track.findAll({
+                where: {
+                    user_id: userId
+                },
+                attributes: ['id','title', 'cover', 'artist', 'preview', "user_id"]
+            });
+            res.status(200).json({
+                message: "Favorite Tracks",
+                data: tracks
+            });
+        } catch (error) {
+            res.status(404).json({
+                message: error.message
+            });
+        }
+    } else{
+        res.status(403).json({
+            message: "Forbidden User."
+        })
     }
 }
 
-// Add authorization
 const deleteFavoriteTrack = async (req,res) => {
-    const userId = req.params.userId;
-    const trackId = req.params.trackId; 
-    try {
-        await Favorite_Track.destroy({
-            where: {
-                id: trackId,
-                user_id: userId
-            }
-        });
-        res.status(200).send({
-            message: `Remove Favorite Track.`
-        });
-    } catch (error) {
-        res.status(404).json({
-            message: error.message
-        });
+    const userId = Number(req.params.userId);
+    const trackId = Number(req.params.trackId); 
+    if(req.user.id === userId){
+        try {
+            await Favorite_Track.destroy({
+                where: {
+                    id: trackId,
+                    user_id: userId
+                }
+            });
+            res.status(200).send({
+                message: `Remove Favorite Track.`
+            });
+        } catch (error) {
+            res.status(404).json({
+                message: error.message
+            });
+        }
+    } else {
+        res.status(403).json({
+            message: "Forbidden User."
+        })
     }
 }
 
-// Add authorization
 const addFollowedArtist = async (req,res) => {
     const body = req.body;
     const name = body.name;
     const picture = body.picture;
-    const userId = req.params.userId;
+    const userId = Number(req.params.userId);
     const artist = {
         name: name,
         picture: picture,
         user_id: userId
     };
-    try {
-        await Followed_Artist.create(artist);
-        res.status(200).json({
-            message: "Add Followed Artist",
-            data: {
-                artist
-            }
-        });
-    } catch (error) {
-        res.status(404).json({
-            message: error.message
-        });
+    if(req.user.id === userId){
+        try {
+            await Followed_Artist.create(artist);
+            res.status(200).json({
+                message: "Add Followed Artist",
+                data: artist
+            });
+        } catch (error) {
+            res.status(404).json({
+                message: error.message
+            });
+        }
+    } else {
+        res.status(403).json({
+            message: "Forbidden User."
+        })
     }
 }
 
-// Add authorization
 const followedArtists = async (req,res) => {
-    const userId = req.params.userId;
-    try {
-        const artists = await Followed_Artist.findAll({
-            where: {
-                "user_id": userId
-            },
-            attributes: ['id', 'name', 'picture']
-        });
-        res.status(200).json({
-            message: "Followed Artists",
-            data: artists
-        });
-    } catch (error) {
-        res.status(404).json({
-            message: error.message
-        });
+    const userId = Number(req.params.userId);
+    if(req.user.id === userId){
+        try {
+            const artists = await Followed_Artist.findAll({
+                where: {
+                    "user_id": userId
+                },
+                attributes: ['id', 'name', 'picture', "user_id"]
+            });
+            res.status(200).json({
+                message: "Followed Artists",
+                data: artists
+            });
+        } catch (error) {
+            res.status(404).json({
+                message: error.message
+            });
+        }
+    } else {
+        res.status(403).json({
+            message: "Forbidden User."
+        })
     }
 }
 
-// Add authorization
 const deleteFollowedArtist = async (req,res) => {
-    const userId = req.params.userId;
+    const userId = Number(req.params.userId);
     const artistId = req.params.artistId; 
-    try {
-        await Followed_Artist.destroy({
-            where: {
-                id: artistId,
-                user_id: userId
-            }
-        });
-        res.status(200).send({
-            message: "Remove Followed Artist."
-        });
-    } catch (error) {
-        res.status(404).json({
-            message: error.message
-        });
+    if(req.user.id === userId){
+        try {
+            await Followed_Artist.destroy({
+                where: {
+                    id: artistId,
+                    user_id: userId
+                }
+            });
+            res.status(200).send({
+                message: "Remove Followed Artist."
+            });
+        } catch (error) {
+            res.status(404).json({
+                message: error.message
+            });
+        }
     }
 }
 
@@ -267,6 +362,7 @@ module.exports = {
     allUsers,
     userProfile,
     registerUser,
+    loginUser,
     editUserProfile,
     deleteUser,
     addFavoriteTrack,
